Move service functions from csr_table_impl.hpp to csr_kernels.hpp
