+++ cpp/oneapi/dal/table/backend/csr_kernels.cpp
@@ -17,6 +17,9 @@
 #include "oneapi/dal/table/backend/csr_kernels.hpp"
 #include "oneapi/dal/table/backend/convert.hpp"
 #include "oneapi/dal/backend/common.hpp"
+#include "oneapi/dal/backend/memory.hpp"
+
+#include <algorithm>
 
 namespace oneapi::dal::backend {
 
@@ -322,6 +325,165 @@ void csr_pull_block(const Policy& policy,
     }
 }
 
+std::int64_t csr_get_non_zero_count(const detail::default_host_policy& policy,
+                                    const std::int64_t row_count,
+                                    const std::int64_t* row_offsets) {
+    if (row_count == 0)
+        return 0;
+
+    return (row_offsets[row_count] - row_offsets[0]);
+}
+
+#ifdef ONEDAL_DATA_PARALLEL
+
+std::int64_t csr_get_non_zero_count(const detail::data_parallel_policy& policy,
+                                        const std::int64_t row_count,
+                                        const std::int64_t* row_offsets) {
+    if (row_count == 0)
+        return 0;
+
+    auto q = policy.get_queue();
+    std::int64_t first_row_offset{ 0L };
+    std::int64_t last_row_offset{ 0L };
+    copy_usm2host(q, &first_row_offset, row_offsets, 1, {}).wait_and_throw();
+    copy_usm2host(q, &last_row_offset, row_offsets + row_count, 1, {}).wait_and_throw();
+
+    return (last_row_offset - first_row_offset);
+}
+
+#endif
+
+/// This function dispatches the execution:
+///     If data parallel policy is enabled and there is a sycl queue associated with row offsets array,
+///         then DPC++ implementation of the function is called.
+///     Otherwise, C++ implementation is called.
+/// This API is needed by ``csr_table_impl`` constructor.
+std::int64_t csr_get_non_zero_count(const array<std::int64_t>& row_offsets) {
+    const std::int64_t row_count = row_offsets.get_count() - 1;
+#ifdef ONEDAL_DATA_PARALLEL
+    const auto optional_queue = row_offsets.get_queue();
+    if (optional_queue) {
+        return csr_get_non_zero_count(
+            detail::data_parallel_policy{ optional_queue.value() },
+            row_count,
+            row_offsets.get_data());
+    }
+#endif
+    return csr_get_non_zero_count(detail::default_host_policy{},
+                                  row_count,
+                                  row_offsets.get_data());
+}
+
+#ifdef ONEDAL_DATA_PARALLEL
+
+template <typename T>
+bool is_sorted(sycl::queue& queue, const std::int64_t count, const T* data) {
+    bool result = true;
+    sycl::buffer<bool, 1> result_buf(&result, sycl::range<1>(1));
+    auto event = queue.submit([&](sycl::handler& cgh) {
+        auto write_result = result_buf.get_access<sycl::access::mode::read_write>(cgh);
+        cgh.parallel_for<struct sorted_checker>(sycl::range<1>(count - 1),
+                                                [=](sycl::id<1> idx) {
+                                                    const int i = idx[0];
+                                                    if (data[i] > data[i + 1]) {
+                                                        write_result[0] = false;
+                                                    }
+                                                });
+    });
+    event.wait_and_throw();
+    return result;
+}
+
+#endif
+
+template <typename T>
+bool is_sorted(const array<T>& arr) {
+    const T* data = arr.get_data();
+    const std::int64_t count = arr.get_count();
+#ifdef ONEDAL_DATA_PARALLEL
+    const auto optional_queue = arr.get_queue();
+    if (optional_queue) {
+        sycl::queue q = optional_queue.value();
+        return is_sorted(q, count, data);
+    }
+#endif
+    return std::is_sorted(data, data + count);
+}
+
+template <typename T>
+out_of_bound_type check_bounds(const std::int64_t count,
+                               const T* data,
+                               const T& min_value,
+                               const T& max_value) {
+    out_of_bound_type result{ out_of_bound_type::within_bounds };
+    for (std::int64_t i = 0; i < count; ++i) {
+        if (data[i] < min_value) {
+            result = out_of_bound_type::less_than_min;
+            break;
+        }
+        if (data[i] > max_value) {
+            result = out_of_bound_type::greater_than_max;
+            break;
+        }
+    }
+    return result;
+}
+
+#ifdef ONEDAL_DATA_PARALLEL
+
+template <typename T>
+out_of_bound_type check_bounds(sycl::queue& queue,
+                               const std::int64_t count,
+                               const T* data,
+                               const T& min_value,
+                               const T& max_value) {
+    int result{ 0 /* out_of_bound_type::within_bounds */ };
+    sycl::buffer<int, 1> result_buf(&result, sycl::range<1>(1));
+    auto event = queue.submit([&](sycl::handler& cgh) {
+        auto write_result = result_buf.get_access<sycl::access::mode::read_write>(cgh);
+        cgh.parallel_for<struct bounds_checker>(
+            sycl::range<1>(count - 1),
+            [=](sycl::id<1> idx) {
+                const int i = idx[0];
+                if (data[i] < min_value) {
+                    write_result[0] = -1 /* out_of_bound_type::less_than_min */;
+                }
+                if (data[i] > max_value) {
+                    write_result[0] = 1 /* out_of_bound_type::greater_than_max */;
+                }
+            });
+    });
+    event.wait_and_throw();
+    out_of_bound_type enum_result{ out_of_bound_type::within_bounds };
+    switch (result) {
+        case -1: enum_result = out_of_bound_type::less_than_min; break;
+        case 0: enum_result = out_of_bound_type::within_bounds; break;
+        case 1: enum_result = out_of_bound_type::greater_than_max; break;
+        default:
+            /* TODO: Error handling */
+            break;
+    }
+    return enum_result;
+}
+
+#endif
+
+template <typename T>
+out_of_bound_type check_bounds(const array<T>& arr,
+                               const T& min_value,
+                               const T& max_value) {
+    const T* data = arr.get_data();
+    const std::int64_t count = arr.get_count();
+#ifdef ONEDAL_DATA_PARALLEL
+    const auto optional_queue = arr.get_queue();
+    if (optional_queue) {
+        sycl::queue q = optional_queue.value();
+        return check_bounds(q, count, data, min_value, max_value);
+    }
+#endif
+    return check_bounds(count, data, min_value, max_value);
+}
+
 #define INSTANTIATE(Policy, BlockData)                                             \
     template void csr_pull_block(const Policy& policy,                             \
                                  const csr_info& origin_info,                      \
@@ -340,5 +502,25 @@ void csr_pull_block(const Policy& policy,
 INSTANTIATE_HOST_POLICY(float)
 INSTANTIATE_HOST_POLICY(double)
 INSTANTIATE_HOST_POLICY(std::int32_t)
+
+#ifdef ONEDAL_DATA_PARALLEL
+
+template bool is_sorted(sycl::queue& queue, const std::int64_t count, const std::int64_t* data);
+template out_of_bound_type check_bounds(sycl::queue& queue,
+                                        const std::int64_t count,
+                                        const std::int64_t* data,
+                                        const std::int64_t& min_value,
+                                        const std::int64_t& max_value);
+
+#endif
+
+template bool is_sorted(const array<std::int64_t>& arr);
+template out_of_bound_type check_bounds(const std::int64_t count,
+                                        const std::int64_t* data,
+                                        const std::int64_t& min_value,
+                                        const std::int64_t& max_value);
+template out_of_bound_type check_bounds(const array<std::int64_t>& arr,
+                                        const std::int64_t& min_value,
+                                        const std::int64_t& max_value);
 
 } // namespace oneapi::dal::backend
